+++ src/main/java/kr/co/easystock/domain/Item/ItemRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item, Integer>, JpaSpecificationExecutor<Item>
+public interface ItemRepository extends JpaRepository<Item, Long>, JpaSpecificationExecutor<Item>
 {
     List<Item> findAllByRetailer(Retailer retailer, Pageable pageable);
 }
