+++ src/main/java/kr/co/easystock/domain/answer/AnswerRepository.java
@@ -3,6 +3,6 @@ package kr.co.easystock.domain.answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface AnswerRepository extends JpaRepository<Answer, Integer>, JpaSpecificationExecutor<Answer>
+public interface AnswerRepository extends JpaRepository<Answer, Long>, JpaSpecificationExecutor<Answer>
 {
 }
