+++ src/main/java/kr/co/easystock/domain/notice/NoticeRepository.java
@@ -3,6 +3,6 @@ package kr.co.easystock.domain.notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface NoticeRepository extends JpaRepository<Notice, Integer>, JpaSpecificationExecutor<Notice>
+public interface NoticeRepository extends JpaRepository<Notice, Long>, JpaSpecificationExecutor<Notice>
 {
 }
