+++ src/main/java/kr/co/easystock/domain/stock/StockRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import java.util.List;
 
-public interface StockRepository extends JpaRepository<Stock, Integer>, JpaSpecificationExecutor<Stock>
+public interface StockRepository extends JpaRepository<Stock, Long>, JpaSpecificationExecutor<Stock>
 {
     List<Stock> findAllByUser(User user);
 }
